fix(sinsp): Try /images/json (without ?digests=1) for podman
Apparently at least the RHEL9 version of podman doesn't properly
respond to /images/json?digests=1, while it does return all
the info we need without the query parameter. Since ?digests=1
is defined in the Docker API, prefer this as the default, but
also try the podman variant.

Note: the API does not return an HTTP error but instead an empty
200 response, so checking the HTTP status is not enough.

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>